Copy parent UI object layer to new child object
Getting dropdowns to work proved to be quite difficult due to the main
canvas being on the UI layer and the camera culling set to UI-only, but
a freshly placed dropdown had the template on the default layer, and
Dropdown creates new canvases for the dropdown list (from the template)
and the blocker. Copying the layer as child objects are created ensures
the template is on the correct layer, but not much can be done about the
blocker (but it doesn't seem to matter).