Rework the skin/style system to be much cleaner
It is both simpler and more powerful (for one, it actually works).

Top level objects (windows, menus, etc) "must" have a skin (otherwise
the default skin is used). Deeper objects may have a skin, in which case
the deeper skin takes over for objects deeper still.

For finding styles within a skin, a style path is built up from the
current object to the first parent with a skin (paths are always
skin-relative). Note that Layout (and derivatives), UIEmpty, and
Viewport do not contribute to a style path except for when searching for
their own style (simplifies scroll views and uses of layouts).