Create a VertiLinkLayoutGroup for flowed text
TMPro supports linked text components such that if the text in the first
component is too long, it flows into the next. This works nicely but not
so well with autolayout. This gets autolayout working for the most part
(TMPro might have a bug in that it seems to use the full text for
calculating the preferred heights of linked text components rather than
only the visible text).

This implementation is a bit brutish. Later iterations might refine
things.