Add a TreeView UI component
The implementation goes for memory efficiency over speed in that tree
view items are created for only those items that are meant to be
visible. This means that when an item is opened, new items are created,
and items are deleted when a parent item is closed. However, game
objects are created or destroyed only when necessary: existing game
objects are reused as much as possible, simply resetting the item to
which they refer.

A bit more work is needed, mainly for item selectabiliy and an action
for when an item is selected.